Moved the camera handling and animation to a separate file.
